+++ packages/preview-middleware-client/test/__mock__/sap/ui/rta/RuntimeAuthoring.ts
@@ -1,0 +1,9 @@
+import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
+
+const RuntimeAuthoringMock = {
+    getService: jest.fn(),
+    getCommandStack: jest.fn(),
+    getFlexSettings: jest.fn()
+};
+
+export default RuntimeAuthoringMock as unknown as RuntimeAuthoring & typeof RuntimeAuthoringMock;