+++ packages/preview-middleware-client/types/sap.ui.rta.d.ts
@@ -1,7 +1,8 @@
 declare module 'sap/ui/rta/command/BaseCommand' {
     import type Element from 'sap/ui/core/Element';
+    import type ManagedObject from 'sap/ui/base/ManagedObject';
 
-    interface BaseCommand {
+   interface BaseCommand extends ManagedObject {
         execute(): Promise<void>;
         getElement(): Element;
     }
@@ -36,22 +37,9 @@ declare module 'sap/ui/rta/command/CommandFactory' {
     import type BaseCommand from 'sap/ui/rta/command/BaseCommand';
     import type DesignTimeMetadata from 'sap/ui/dt/DesignTimeMetadata';
     import type Element from 'sap/ui/core/Element';
-    import type { Layer } from 'sap/ui/fl';
-
-    export interface FlexSettings {
-        layer: Layer;
-        developerMode: boolean;
-        baseId?: string;
-        projectId?: string;
-        scenario?: string;
-        namespace?: string;
-        rootNamespace?: string;
-    }
+    import type { FlexSettings } from 'sap/ui/rta/RuntimeAuthoring';
 
     interface CommandFactory {
-        /**
-         *
-         */
         getCommandFor<T extends BaseCommand = BaseCommand>(
             control: Element | string,
             commandType: string,
@@ -78,14 +66,8 @@ declare module 'sap/ui/rta/command/OutlineService' {
         icon?: string;
     }
 
-    /**
-     *
-     */
     interface OutlineService {
         get(): Promise<OutlineViewNode[]>;
-        /**
-         *
-         */
         attachEvent<T>(eventName: string, handler: (params: T) => void): void;
     }
 
@@ -97,16 +79,44 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
     import type Stack from 'sap/ui/rta/command/Stack';
     import type ElementOverlay from 'sap/ui/dt/ElementOverlay';
     import type ContextMenu from 'sap/ui/dt/plugin/ContextMenu';
+    import type { Layer } from 'sap/ui/fl';
 
-    interface RuntimeAuthoring {
-        attachSelectionChange(handler: (event: Event) => void): void;
+    export interface FlexSettings {
+        layer: Layer;
+        developerMode: boolean;
+        baseId?: string;
+        projectId?: string;
+        scenario?: string;
+        namespace?: string;
+        rootNamespace?: string;
+    }
+
+    export type SelectionChangeEvent = Event<SelectionChangeParams>;
+    export interface SelectionChangeParams {
+        selection: ElementOverlay[];
+    }
+
+   interface RuntimeAuthoring {
+        attachSelectionChange(handler: (event: SelectionChangeEvent) => void): void;
         attachModeChanged: (handler: (event: Event) => void) => void;
         attachUndoRedoStackModified: (handler: (event: Event) => void) => void;
         getCommandStack: () => Stack;
         getService: <T>(name: 'outline' | string) => Promise<T>;
         getSelection: () => ElementOverlay[];
         getDefaultPlugins: () => { contextMenu: ContextMenu };
+        getFlexSettings: () => FlexSettings
     }
 
     export default RuntimeAuthoring;
+}
+
+declare module 'sap/ui/rta/api/startAdaptation' {
+    import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
+    
+    export type RTAPlugin = (rta: RuntimeAuthoring) => void;
+    export type StartAdaptation = (options: object, plugin?: RTAPlugin) => void;
+
+    const startAdaptation: StartAdaptation;
+
+    export default startAdaptation;
 }