+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/ui-components
 
+## 1.11.6
+
+### Patch Changes
+
+-   50bf49ac: chore(deps): update storybook to v7.4.1
+
+## 1.11.5
+
+### Patch Changes
+
+-   30fc4a11: feat(ui-components) add createref as props to UICreateSelect
+
 ## 1.11.4
 
 ### Patch Changes